+++ src/Networking/UDPMessage.h
@@ -2,13 +2,14 @@
 #define udpmessage_h
 #include <codecvt>
 #include <SDL2_net/SDL_net.h>
+#include "Utilities/helperFunctions.h"
 
 class UDPMessage// : Singleton<UDPMessage>
 {
 	public:
 		UDPMessage(IPaddress ip);
 		void SendUnreliableMessage(Uint8 data);
-		static IPaddress StringToIP(std::string stringIP, int port);
+		static IPaddress StringToIP(std::string stringIP, int16 port);
 		bool ReceiveMessage(Uint8& outValue);
 	private:
 		UDPsocket socket;
